refactor(units): SI time header (#4)
